requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asteval==1.0.5
attrs==23.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
build==1.0.3
CacheControl==0.13.1
cachetools==5.5.0
Cerberus==1.3.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.16
circuitbreaker==2.0.0
cleo==2.1.0
click==8.1.7
cloudpickle==3.1.0
coloredlogs==15.0.1
contourpy==1.3.0
crashtest==0.4.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.4
datasets==3.0.2
decorator==4.4.2
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
durationpy==0.9
faiss-cpu==1.9.0
fastapi==0.115.4
fastjsonschema==2.19.1
ffmpy==0.4.0
filelock==3.13.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
googleapis-common-protos==1.65.0
gradio==5.4.0
gradio_client==1.4.2
greenlet==3.0.3
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.6
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib-metadata==7.0.1
importlib_resources==6.4.5
inflection==0.5.1
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.3.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==24.3.0
kiwisolver==1.4.7
kubernetes==31.0.0
lancedb==0.14.0
langchain==0.3.4
langchain-chroma==0.1.4
langchain-community==0.3.3
langchain-core==0.3.13
langchain-openai==0.2.4
langchain-text-splitters==0.3.0
langsmith==0.1.137
llama-index==0.11.0
llama-index-agent-openai==0.3.0
llama-index-cli==0.3.0
llama-index-core==0.11.0.post1
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-langchain==0.2.1
llama-index-embeddings-openai==0.2.0
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48
llama-index-llms-openai==0.2.0
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.0
llama-index-readers-llama-parse==0.2.0
llama-parse==0.4.0
llamaindex-py-client==0.1.13
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.12.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
nodemon-py-simple==0.0.7
numpy==1.26.4
oauthlib==3.2.2
oci==2.137.1
ocifs==1.3.1
onnxruntime==1.19.2
openai==1.52.2
opencv-python==4.10.0.84
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
oracle_ads==2.12.2
orjson==3.10.10
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
posthog==3.7.0
predictionguard==2.4.0
proglog==0.1.10
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
py==1.11.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pylance==0.18.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jsonschema-objects==0.5.5
python-multipart==0.0.12
pytubefix==8.2.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
ruff==0.7.1
safehttpx==0.1.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentence-transformers==3.2.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.41.2
streamlit==1.39.0
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tantivy==0.22.0
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
tomlkit==0.12.0
torch==2.5.1
tornado==6.4.1
tqdm==4.66.6
transformers==4.46.1
trove-classifiers==2024.1.8
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.25.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
wrapt==1.16.0
xattr==0.10.1
xxhash==3.5.0
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.17.0
